starts restructuring to implement multiple phases of rule application
